Change auto shoot formula back to `ShooterFormula`.
