+++ src/main/java/frc/robot/commands/AutoShootCommand.java
@@ -2,11 +2,11 @@ package frc.robot.commands;
 
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.wpilibj2.command.Command;
+import frc.robot.constants.ShooterConstants;
 import frc.robot.subsystems.Acquisition;
 import frc.robot.subsystems.DriveTrain;
 import frc.robot.subsystems.Kicker;
 import frc.robot.subsystems.Shooter;
-import frc.robot.utils.AutoShootUtils;
 
 public class AutoShootCommand extends Command {
     private final Acquisition acquisition = Acquisition.getInstance();
@@ -55,7 +55,7 @@ public class AutoShootCommand extends Command {
     private void autoShoot(double nearShooterHubDistance, double farShooterHubDistance) {
         double minDistance = Math.min(nearShooterHubDistance, farShooterHubDistance);
         double maxDistance = Math.max(nearShooterHubDistance, farShooterHubDistance);
-        for (AutoShootUtils.AutoShootFormula formula : AutoShootUtils.AUTO_SHOOT_FORMULAS) {
+        for (ShooterConstants.ShooterFormula formula : ShooterConstants.SHOOTER_FORMULAS) {
             if (formula.getMin() <= minDistance && formula.getMax() >= maxDistance) {
                 shooter.getNearShooter().setAngle(formula.getAngle());
                 shooter.getNearShooter().start(formula.getShooterRPM(nearShooterHubDistance));
